move source code generation to separate file
